requirements.txt
aiohttp
aiosqlite
discord.py
ffmpeg